requirements.txt
mediapipe==0.10.14
opencv-python>=4.10.0
numpy>=1.26.0
websockets>=12.0
spotipy>=2.24.0
python-dotenv>=1.0.1
